requirements.txt
fastapi
uvicorn
python-multipart
pillow
torch
torchvision
timm
pydantic
onnx
onnxruntime-gpu
python-dotenv
